// src/app/models/interfaces/IMasterData.ts
export interface IMasterData {
  buildings: any[];
  columns: any[];
  costCenters: any[];
  costCodes: any[];
  countries: any[];
  cities: any[];
  criticalities: any[];
  depreciationMethods: any[];
  existingAssets: any[];
  financeCodes: any[];
  floors: any[];
  rooms: any[];
  masterDevelopments: any[];
  maintainabilityTypes: any[];
  parentAssets: any[];
  states: any[];
  subSystemCodes: any[];
  subSystemParameters: any[];
  suppliers: any[];
  subContractors: any[];
  systemCodes: any[];
  tagCodes: any[];
  typeCodes: any[];
  unitsORSpaces: any[];
  uomCodes: any[];
  zones: any[];
}